backend/requirements.txt
appnope==0.1.0
asgiref==3.2.3
aspy.yaml==1.3.0
astroid==2.3.3
backcall==0.1.0
bandit==1.6.2
boto3==1.11.9
botocore==1.14.9
certifi==2019.11.28
cfgv==2.0.1
chardet==3.0.4
Click==7.0
coverage==5.0.3
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.2
django-allauth==0.41.0
django-model-utils==4.0.0
django-naomi==0.8
django-storages==1.8
django-widget-tweaks==1.4.5
docutils==0.15.2
dodgy==0.2.1
dparse==0.4.1
gitdb2==2.0.6
GitPython==3.0.5
gunicorn==20.0.4
identify==1.4.11
idna==2.8
importlib-metadata==1.4.0
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
jmespath==0.9.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
model-mommy==2.0.0
nodeenv==1.3.4
oauthlib==3.1.0
packaging==20.1
parso==0.6.0
pbr==5.4.4
pep8-naming==0.4.1
pexpect==4.8.0
pickleshare==0.7.5
pre-commit==1.21.0
prompt-toolkit==3.0.3
prospector==1.2.0
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycodestyle==2.4.0
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pylint-celery==0.3
pylint-django==2.0.12
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.6
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
requirements-detector==0.6
s3transfer==0.3.2
safety==1.8.5
setoptconf==0.2.0
six==1.14.0
smmap2==2.0.5
snowballstemmer==2.0.0
sqlparse==0.3.0
stevedore==1.31.0
stripe==2.42.0
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==16.7.9
wcwidth==0.1.8
whitenoise==5.0.1
wrapt==1.11.2
zipp==2.1.0
Pillow==7.0.0
